Fix warnings related to dangling comments and lossy conversions
Recently numerous warnings have displayed related to lossy conversions
and dangling `javadoc` comments.

This update ensures that there are no dangling `javadoc` comments these
changes fall into various categories such as unused code that has
`javadoc` comments embedded, empty `javadoc` comments, duplicate
`javadoc` comments, or javadoc comments that are not located above a
declaration.

Updates were also made to `XDHPublicKeyImpl`, `BaseTestChaCha20KAT`,
and `BaseTestChaCha20NoReuse` to explicitly cast an int to a byte OR
make use of a common utility function `BaseUtils.hexStringToByteArray`
that does not have such a lossy conversion present.

While fixing dangling comments one test `testRSASignatureSHA512_256` was
found to be commented out. It appears to execute correctly and does seem
like a valid test so this test was uncommented.

Two spaces found in the pom.xml were removed for consistency.

+++ src/main/java/com/ibm/crypto/plus/provider/XDHPublicKeyImpl.java
@@ -61,7 +61,7 @@ final class XDHPublicKeyImpl extends X509Key implements XECPublicKey, Destroyabl
         int bMod8 = (keyArray.length * 8) % 8;
         if (bMod8 != 0) {
             int msk = (1 << bMod8) - 1;
-            keyArray[0] &= msk;
+            keyArray[0] &= (byte)msk;
         }
 
         this.u = new BigInteger(1, keyArray); // u is the public key reversed
@@ -123,7 +123,7 @@ final class XDHPublicKeyImpl extends X509Key implements XECPublicKey, Destroyabl
             int bMod8 = (reverseKey.length * 8) % 8;
             if (bMod8 != 0) {
                 int msk = (1 << bMod8) - 1;
-                reverseKey[0] &= msk;
+                reverseKey[0] &= (byte)msk;
             }
 
             this.u = new BigInteger(1, reverseKey); // u is the public key reversed
